ransomeye_posture_engine/requirements.txt
# Path and File Name : /home/ransomeye/rebuild/ransomeye_posture_engine/requirements.txt
# Author: nXxBku0CKFAJCBN3X1g3bQk7OxYQylg8CMw1iGsq7gU
# Details of functionality of this file: Python dependencies for RansomEye Posture Engine

# Database
asyncpg>=0.29.0

# YAML parsing
PyYAML>=6.0

# Cryptography (for signing)
cryptography>=41.0.0

# PDF generation (optional - falls back to text if not available)
reportlab>=4.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

